requirements.txt
langchain-nomic
langchain_community
langchain-ollama
langchain-chroma
tiktoken
langchainhub
chromadb
langchain
langgraph
tavily-python
gpt4all
firecrawl-py
python-dotenv